+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v1.1.0][v1.1.0] - 30-Oct-2021
+
+### Milestone: [catapult-server@v1.0.2.0][catapult-server@v1.0.2.0]
+
+Package  | Version  | Link
+---|---|---
+REST Core| v2.3.7 | [catapult-rest][catapult-rest@v2.3.7]
+SDK Core| v1.0.1 | [symbol-sdk][symbol-sdk@v1.0.1]
+
+- support https/http nodes. [#64](https://github.com/symbol/statistics-service/pull/64)
+- Removed node rewards related source code. [#65](https://github.com/symbol/statistics-service/issues/65)
+- Added new filter on nodes endpoint [#71](https://github.com/symbol/statistics-service/issues/71)
+
 ## [v1.0.1][v1.0.1] - 25-May-2021
 
 ### Milestone: [catapult-server@v1.0.0.0][catapult-server@v1.0.0.0]
@@ -86,6 +99,8 @@ REST Core| v2.1.0 | [catapult-rest](https://github.com/nemtech/catapult-rest/rel
 
 ### Fixes
 - Cors error. [#13](https://github.com/nemgrouplimited/symbol-statistics-service/issues/13)
+
+[v1.1.0]: https://github.com/symbol/statistics-service/releases/tag/v1.0.1
 
 [v1.0.1]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v1.0.1
 [v1.0.0]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v1.0.0